+++ src/current/current_runner.py
@@ -118,13 +118,13 @@ def get_n_save_whole_year_runner_data() -> set:
 
 
     df = df.replace("-",value=np.nan)
-    df['year'] = df['year'].astype(int)
-    df['id'] = df['id'].astype(int)
-    df['G'] = df['G'].astype(int)
-    df['SB'] = df['SB'].astype(int)
-    df['CS'] = df['CS'].astype(int)
-    df['OOB'] = df['OOB'].astype(int)
-    df['PKO'] = df['PKO'].astype(int)
+    df['year'] = df['year'].astype('int32')
+    df['id'] = df['id'].astype('int32')
+    df['G'] = df['G'].astype('int32')
+    df['SB'] = df['SB'].astype('int32')
+    df['CS'] = df['CS'].astype('int32')
+    df['OOB'] = df['OOB'].astype('int32')
+    df['PKO'] = df['PKO'].astype('int32')
 
     save_df(
         df,
